# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Projeto de PDWA5",
        "description": "API do projeto de PDWA5",
        "contact": {
            "email": "admin@admin.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost/api",
            "description": "Endpoint da API."
        }
    ],
    "paths": {
        "/auth": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Autenticar e gerar token",
                "description": "Retorna token de autenticação",
                "operationId": "e2e910d793673bfad04a80eda7ad070d",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "User Name"
                                    },
                                    "pwd": {
                                        "type": "string",
                                        "example": "12345#qwert"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Token gerado com sucesso.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "token": {
                                            "type": "string",
                                            "example": "1|eoiuywourioeuwoUIymlajsklaOPImewruwopruokhdshjfdf"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Ao falhar na autorização."
                    },
                    "422": {
                        "description": "Ao falhar em uma validação.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "name": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "": {
                                                        "type": "string"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Auth"
                ],
                "summary": "Revogar token",
                "description": "Revoga token de autenticação",
                "operationId": "df2eee26444444a49a587c8073a4bf74",
                "responses": {
                    "200": {
                        "description": "Token gerado com sucesso."
                    },
                    "401": {
                        "description": "Ao falhar na autorização."
                    }
                }
            }
        },
...
    }
}